### What problem does this PR solve?

Issue Number: None

Related PR: #64891

Problem Summary: Merge upstream master at 0d9ffb05f35 into the multi-stage predicate lazy materialization branch. Resolve SegmentIterator execution-path changes while preserving master common-expression handling, retain both FE session-variable tests, and move the unpublished multi-stage Thrift option to field 228 to avoid colliding with master field 227.

### Release note

None

### Check List (For Author)

- Test: Unit Test and Regression Test
    - ASAN FE and BE builds
    - PredicateLmStage1SelectorTest
    - SessionVariablesTest
    - SegmentIteratorLazyPrunedTest
    - correctness_p0/test_multi_stage_predicate_lm
- Behavior changed: No
- Does this need documentation: No

+++ be/test/runtime/runtime_profile_counter_tree_node_test.cpp
@@ -167,10 +167,10 @@ TEST_F(RuntimeProfileCounterTreeNodeTest, HighWaterMarkCounterToThrift) {
     childCounterMap["root"].insert("child");
 
     rootCounter->add(10);
-    rootCounter->set(5);
+    rootCounter->set(int64_t(5));
 
     childCounter->add(100);
-    childCounter->set(50);
+    childCounter->set(int64_t(50));
 
     RuntimeProfileCounterTreeNode rootNode = RuntimeProfileCounterTreeNode::from_map(
             counterMap, childCounterMap, RuntimeProfile::ROOT_COUNTER);
